6. Level Class, Camera centered on player
Added a Level class
Renamed Rectangle class to Block

Level has a vector with Blocks that renders every frame.

Player handles input and updates its x and y.

Player is always rendered in the middle of the screen.

When blocks are rendered, it takes in account the player or "camera"
position and renders the blocks on the right position so that it looks
like everything moved.

Added small randomization of the map

Added a collision function in Level class that is called by the player
during update. The function will detect if the player is about to move
into a block and will make sure player doesnt move there.